Add clientaccesspolicy.xml & crossdomain.xml
Not having these files causes unnecessary 404 errors. They've been written to be as restrictive as possible, nothing is allowed. This is the same as not having them.

See: https://www.blue-bag.com/blog/why-are-you-looking-my-crossdomainxml

We still need to add a valid favicon.ico file to stop the constant lookups for that file.